Extract module for specs of edits that use ValidationStats
Holds shared setup and helper methods to make it easier to test this
kind of edit.
